+++ src/i18n.js
@@ -204,13 +204,16 @@ i18n
             accountMenuProfileAltText: "{{name}}'s account",
             raspberryPiLogoAltText: "Raspberry Pi logo",
           },
+          projectName: {
+            label: "Project name",
+            newProject: "New Project",
+          },
           header: {
             autoSaving: "Saving",
             autoSaved: "Saved",
             download: "Download",
             downloadFileNameDefault: "my {{project_type}} project",
             editorLogoAltText: "Editor logo",
-            newProject: "New Project",
             projects: "Your projects",
             renameProject: "Edit project name",
             renameSave: "Save project name",
@@ -219,6 +222,11 @@ i18n
           },
           imagePanel: {
             gallery: "Image Gallery",
+          },
+          projectsPanel: {
+            projects: "Projects",
+            yourProjectsButton: "Go to your projects",
+            projectTypeLabel: "Project type",
           },
           infoPanel: {
             info: "Information",
@@ -254,9 +262,7 @@ i18n
               inputLabel: "Project name",
             },
             projectType: {
-              html: "HTML",
               inputLabel: "What kind of project do you want to make?",
-              python: "Python",
             },
           },
           notifications: {
@@ -373,6 +379,10 @@ i18n
             python_type: "Python",
             html_type: "HTML",
           },
+          projectTypes: {
+            html: "HTML/CSS",
+            python: "Python",
+          },
           runButton: {
             run: "Run",
             stop: "Stop",
@@ -387,6 +397,7 @@ i18n
             file: "Project files",
             images: "Image gallery",
             settings: "Settings",
+            projects: "Projects",
             information: "Information",
             information_text:
               "Our code editor is designed to create a simple and safe learning environment for beginner programmers. As a result, we have implemented some limitations, such as the inability to include external links in your code.",