+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New landing page (#531)
 - Ability to specify project name and type on creation (#519)
 - Ability to view HTML preview output in separate window (#536)
+- Add Projects Panel to sidebar (#564)
 
 ### Changed